Decide how we should ignore in‐progress casts
s##e##/episode.txt and other files can exist, but the head value of
PUBLISHED needs to be set to make it visible in the Feed, show the
episode in the site list and the notes on the cast page. Otherwise site
viewers just see the file link if they ‘guess’ the URL.
Obviously FILENAME needs to be set too, but that’s obvious!