Made events countdown more pleasant in the backend
Have re-used previous CSS for the timer, but obviously it doesn’t fit
perfectly with the style. This is obviously going to need more tweaking.

Looking at the existing code style of some of these files is painful,
will be happy when #117 is nailed down, as for the bulk sections where
we fill a temporary variable with some HTML? shudder.

See the example at the bottom of the change as a tidy use of HEREDOC to
make these segments readable.

Also resolved that p/div concern.