Use protocol aware URLs when we can
Now all images should be pulled from http(s)://steamcdn-a.akamaihd.net
not entirely sure this is correct, or acceptable, but steamstatic is not
useful for us here.

Renamed img_small to img_icon to better describe what it is, this
filename is hash based, so we rely on Steam giving us the correct image.
Unlike the header and small header, which are just based on the appid.

Corrected an off‐by‐one error that appeared in the bootstrap update.