Add a URL for generating the YouTube image
This produces an SVG so we can have decent quality images for the video,
it current is lacking support for: Guest images and Guest
featured-games, as we do not store this information anywhere yet.

Locally I do have some guest / game deets, but they need to be baked
into the shownote files…

This image still contains the previous gravel background, and some
commented out blocks. Including that for completeness sake, can be
removed in future commits.

We could naïvely include all hosts/guests in each image, it would make
it much more simple, but then we get larger files and your SVG viewer
might load additional resources that are not needed. Aiming for the
future, of course.