Youtube thumbnailer now supports games!
Misc changes all over to fix some behaviours I noticed on staging
after the last push.

For unknown people, it will generate a unique hash so there shouldn’t be
ID collisions in the SVG anymore. Repeated Hosts/Guests would still
cause an issue, but people aren’t going to do that, right?!

Some refactoring was done as I noticed there was repetition, but fully
aware there is still more to do with nameplate() and name() sharing much
of same code.

To avoid having to special case host/guest <defs>, trying something new
with css3 transforms on the each person’s <text>. This may not be
supported by all viewers, but I want to test it a little before deciding
to remove it.

Reverted the forenames being shown on the cast list, improved behaviour
for people without a Twitter handle.