Fix the issue with our Steam group link breaking the menu
This also resolves(*) the issue of the missing menu, as we now have an
element that will expand the menu once it has collapsed.

* I say resolved, but this is reliant on javascript for now. I assume I
would be able to do some :hover :focus :target hack to make it somewhat
work without, but that is if this is tolerable at all

Had to move Projects menu because dropdown causes a scroll in the navbar
list. LOL.