Hint to bootstrap the priority of the columns
Having issues as Recorded header derps around in size due to the added
sorting handlers and padding. Nothing behaves well, been poking for 20
minutes.

Started this because we need more room for guests, as just having their
avatar is not enough to inform the viewer of who they are - even with
making them a clickable link.