Improve layout with no cast event, add RSS icons and improve style
Would say this closes #44, we now have a more compact lead in to the
episodes listing. Remaining work should be on pagination and fixing the
column width glitchy mess bootstrap gives us.

As this is on the WIP branch, you may check the live changes at
https://staging.steamlug.org/cast.php

Icon is from http://www.feedicons.com/ Home of the free standard RSS
Icon /ad