Don’t <a href=""> protocol-naked URLs in notes
To make sure the RSS feed doesn’t break strict validation (as I just
noticed it had done with s02e14), we have to stop being lax about
accepting links in notes. We now require http(s):// for all the scripts
to turn that into a link.

We’d had this occur before with a nordicgames.at link, as the removed
comment suggests.

This is part of making cast + feed share same code.