Make servers more responsive, even if it takes the same time to load
We disable output compression to prevent the httpd buffering output,
then call a flush(); and pray that it has had enough output to consider
sending tp the client.
At least for me locally, this means I see the <title> as soon as I
navigate, rather than having just a spinner for a few seconds.

Moved code around because of the above, fixed a few tests (isset is
rubbish) and restored a typo (oops fa-lock)