Put user sign-in front & centre
This is now pushing user data onto our site, personaname specifically,
that we clean up naïvely before using. Please scan over this change for
mistakes or security issues.

We also now redirect the user on log‐in to a new page, loggedin.php,
which takes a query of the previous page - so we can inform them of what
we do with their Steam information, and give them an option to log out
again.
And if they are not a member of the Steam group, offer a link to it.