Consistently use htmlspecialshars_decode content from Steam
Noticed with the latest news post, SteamLUG Cast S03E03 Now Available
&amp; Ryan &quot;Icculus&quot; Gordon Appearing on S03E04! were being
double HTML encoded, our code was calling htmlspecialchars on the
assumption it /wasn’t/ encoded. Yet further on, our code assumed code
was encoded (for the description).

Since we’re using this content as #text, it is safer to avoid
escapement. Revert if dumb.