Resolve last style issue in #132, panels now all match
I understand this class is meant for channeling the user to look at more
important things, but for a page like our Live Stream, or News, having
one item stand out works the opposite way. For now, all panels are
panel-default.
With this change, I have increased the contrast of the title color with
the header background.