cast.php now uses functions_cast for head parsing
Point of all this work is to factor reuse into chunks. Feed + Cast
scripts have a bunch of repeated processes, and I am worried that at
one point we are going to make a mistake!

nameplate() stuff still needs looking into, as well as looking into
making this function do the HTMLifying and sanity checking?