New helper script for group member playtime stats
This is basically a PHP-isation of @Nemoder’s same script, with a few
small additions… it works, but performance is not great. Commit time to
the db is overly length, and script runtime is … over an hour.

It would be sensible to, now that game stats aren’t reliant on game
info, to break updating games list another task, and maybe batch up
fetching members so if the script times out, we do less repeated work.

ATM, it is configured to require admins to trigger it, but I expect it
will be moved out of tree for the server to run, at that point, we can
retire that requirement.