Add a minified version of our current style
We need to keep this in sync with the un-minified version, and at some
point find a tool that we consistently pump the source through to avoid
excessive diffs on the min file.
This is our last resolvable issue on the Page Speed front for improving
page load times, as remaining cache-able resources are 3rd party, we
wish to avoid the FOUC problem, and tablesorter still requires a chunk
of jQuery. Future TODO is retire those two js for a compact one.