Add a bloody touch icon, URL rewriting for all Apple devices, HTML for Android
Advice from https://mathiasbynens.be/notes/touch-icons we want to offer
this file to stop the 404s, iOS will always look in / without meta
headers, Android wont but we can point them to our standard file with
the HTML5 rel="icon" which is the "new" shortcut reference. We will
still keep the favicon, for legacy browsers.

This image needs to be updated when we actually get a favicon, and that
is #128