Update db schema to latest
Don't have access to desktop to get the actual Poll table schema, so
making it from inference from the code. This obviously needs to be
corrected if anything is widely out of sync. Working on tweaks, as I
would like to expand the functionality of Polls, at least for clan
support.
And the polling code needs a minor clean to sync up with the rest in the
repo now.