Remove both avatar files when admin triggers a delete
When we added resize support, we made the site generate a backup image,
and resize into a new file. Our deletion code did not know about this,
so when you go to upload a new file, it would fail because the file
already exists.

Fixes a bug introduced in 9ed144d932635