Fix our minified CSS to include both sources
To support the site glyphs for links and tables, as well as our event
ticker, we have a small custom font that has a minimal CSS snippet. This
needs to be included in each update of the .min.css file, as we have not
hooked up #192 yet.

The current process is:

```
cat bootstrap.steamlug.css steamlugfont.css | yui-compressor --type css
--line-break 420 -v -o bootstrap.steamlug.min.css
```