Revert "Revert "Update CSS for steam login button""
This reverts commit a8d4c19201d1e6216b9872ca92ad006fa564a802.

The commit message states:

> Thanks for #199 Jason however we concluded the previous button was
better for the site.

Sorry, but I do not agree that "we" came to this conclusion. Two people (@meklu and me) in the steamlug.org group were okay with it, while two others (@johndrinkwater and @Corben78) disputed it after it sat for 9 months and was merged. Nobody else had any comments on the matter. There was no conclusion.

I think it is unfair for @johndrinkwater to decide to make the change without further conversation, which #207 was made to facilitate. As @johndrinkwater has made clear to me before, it is important that people don't just commit to this repo at will, with no review beforehand.