Migrate server flags from images to emoji
Purpose of this is to reduce our reliance on images (allowing us to
expand supported regions without uploading more files to the repo), and
let platforms draw their native flag designs.

Mozilla has been shipping Firefox with the EmojiOne font set for emoji
support for a while, and various mobile platforms and desktops do
support this character block.

This PR is for testing support before we merge it.