Avoid the else branch for memcache use, makes code clearer
When access to memcache is offline, fetchOrStore will still return the
result of the function call; this makes the code more readable and
prevents copy/paste errors for the false branch. This mirrors the
behaviour added for the servers memcache update.

This does remove error reporting for memcache calls… because otherwise
passing in false here would produce errors in log for every visit.